Merge `raisesCondition()` into `raises()`.
`EitherType` attempts to unify the left type first, so it will match `T` if possible, and default to `Any` if not. This only works when `type` comes before `condition`, and since both arguments are
optional, it's possible the user will pass only the function, which will be incorrectly interpreted as being `type`, so we have to check for that. Arguably, the same can be said of `msgNotThrown`.