Fix issues in `Assert.exception()`. (#129)
* Use `@deprecated` JSDoc tag.

This allows the IDE to notice it, without being as disruptive as the `@:deprecated` metadata.

* Rewrite and organize `_raisesImpl()`.

This fixes a null pointer error in the tests, moves the `try catch` block to the top of the function for easier readability, reduces the total amount of code, and standardizes the failure messages. (Among other things, it used to print a double space if `type` was null.)

* Update documentation to use new function name.

* Don't test enum exceptions in C++.

On that target, `Std.isOfType()` will return true as long as `e` is any enum value, making it useless for the intended test.